Validate repo without behaviour check
Repo wrappers that don't implement all functions of the Ecto.Repo
behaviour are still viable and shouldn't be validated with a behaviour
check. This changes repo validation back to the way it was done in older
versions, by checking that it's a valid module that exports `config/0`.

All other `:behaviour` checks are for behavious defined by Oban, and
remain intact.

Closes #1014